Eclipse warning cleanup: primarily unnecessary casts

+++ ptolemy/actor/lib/jai/DoubleMatrixToJAI.java
@@ -175,13 +175,13 @@ public class DoubleMatrixToJAI extends Transformer {
                 break;
 
             case _INT:
-                _maxValue = (double) Integer.MAX_VALUE;
-                _minValue = (double) Integer.MIN_VALUE;
+                _maxValue = Integer.MAX_VALUE;
+                _minValue = Integer.MIN_VALUE;
                 break;
 
             case _SHORT:
-                _maxValue = (double) Short.MAX_VALUE;
-                _minValue = (double) Short.MIN_VALUE;
+                _maxValue = Short.MAX_VALUE;
+                _minValue = Short.MIN_VALUE;
                 break;
 
             case _USHORT:
@@ -190,11 +190,11 @@ public class DoubleMatrixToJAI extends Transformer {
                 break;
 
             case _FLOAT:
-                _maxValue = (double) Float.MAX_VALUE;
+                _maxValue = Float.MAX_VALUE;
                 break;
 
             case _DOUBLE:
-                _maxValue = (double) Double.MAX_VALUE;
+                _maxValue = Double.MAX_VALUE;
                 break;
 
             default:
