+++ ptolemy/copernicus/kernel/fragment/JavaFragmentParserTokenManager.java
@@ -1226,7 +1226,7 @@ public class JavaFragmentParserTokenManager implements
                     }
                 } while (i != startsAt);
             } else {
-                int hiByte = (int) (curChar >> 8);
+                int hiByte = (curChar >> 8);
                 int i1 = hiByte >> 6;
                 long l1 = 1L << (hiByte & 077);
                 int i2 = (curChar & 0xff) >> 6;
@@ -1347,7 +1347,7 @@ public class JavaFragmentParserTokenManager implements
                     }
                 } while (i != startsAt);
             } else {
-                int hiByte = (int) (curChar >> 8);
+                int hiByte = curChar >> 8;
                 int i1 = hiByte >> 6;
                 long l1 = 1L << (hiByte & 077);
                 int i2 = (curChar & 0xff) >> 6;
